server/requirements.txt
asgiref==3.5.2
django==4.1.3
sqlparse==0.4.3
djangorestframework==3.14.0
pytz==2022.6
django-cors-headers==3.13.0
Pillow==9.3.0
djangorestframework-simplejwt==5.2.2
pyjwt==2.6.0